// client/src/store/auth/types.ts
export enum AUTH_ACTION_TYPES {
	SET_ACCESS_TOKEN = "auth/SET_ACCESS_TOKEN",
	SET_REFRESH_TOKEN = "auth/SET_REFRESH_TOKEN",

	LOGIN_START = "auth/LOGIN_START",
	LOGIN_FAILURE = "auth/LOGIN_FAILURE",
	SET_USER = "auth/SET_USER",
	SET_TEMPORARY_PERMISSIONS = "auth/SET_TEMPORARY_PERMISSIONS",
	CLEAR_TEMPORARY_PERMISSIONS = "auth/CLEAR_TEMPORARY_PERMISSIONS",

	LOGOUT_START = "auth/LOGOUT_START",
	LOGOUT_FAILURE = "auth/LOGOUT_FAILURE",
	LOGOUT_SUCCESS = "auth/LOGOUT_SUCCESS",
	CLEAR_AUTH_ERROR = "auth/CLEAR_AUTH_ERROR",
	CLEAR_AUTH_LOADING = "auth/CLEAR_AUTH_LOADING",

	// Auto Logout mechanism
	USER_ACTIVITY_DETECTED = "auth/USER_ACTIVITY_DETECTED",
	REMOTE_USER_ACTIVITY_DETECTED = "auth/REMOTE_USER_ACTIVITY_DETECTED",
	SHOW_LOGOUT_WARNING = "auth/SHOW_LOGOUT_WARNING",
	HIDE_LOGOUT_WARNING = "auth/HIDE_LOGOUT_WARNING",
	CONFIRM_LOGOUT = "auth/CONFIRM_LOGOUT",
	CANCEL_LOGOUT = "auth/CANCEL_LOGOUT",
	REFRESH_TOKENS_START = "auth/REFRESH_TOKENS_START",
	REFRESH_TOKENS_SUCCESS = "auth/REFRESH_TOKENS_SUCCESS",
	REFRESH_TOKENS_FAILURE = "auth/REFRESH_TOKENS_FAILURE",
	SET_INACTIVITY_TIMEOUT = "auth/SET_INACTIVITY_TIMEOUT",
	SET_ENTITY_OBJ = "auth/SET_ENTITY_OBJ",
	SET_SELECTED_ACCOUNT_IDS = "auth/SET_SELECTED_ACCOUNT_IDS",
	SET_SCHOOL = "auth/SET_SCHOOL",
}

export type SessionLogoutPayload = {
	code: "SESSION_EXPIRED" | "SESSION_REVOKED" | "SESSION_TIMED_OUT" | null;
	message: string;
};
export enum SESSION_LOGOUT_CODE {
	SESSION_EXPIRED = "SESSION_EXPIRED",
	SESSION_REVOKED = "SESSION_REVOKED",
	SESSION_TIMED_OUT = "SESSION_TIMED_OUT",
}